requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
dataclasses==0.8
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==1.0.0
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.1
lark-parser==0.12.0
MarkupSafe==2.0.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
required==0.4.0
six==1.16.0
typing-extensions==4.0.1
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
zipp==3.6.0
